Wire PCN↔batch CaseLinks with ExUI-resolvable link reasons.
Own links on the batch caseLinks collection, emit CLRC007 + OtherDescription,
and stub CaseLinkingReasonCode in the local ExUI nav proxy so Reasons render.

+++ src/main/java/uk/gov/hmcts/reform/tecpoc/ccd/EnforcementCaseConfiguration.java
@@ -176,7 +176,7 @@ public class EnforcementCaseConfiguration
                 );
             }
             Long existingEnforcement = repository.findEnforcementCaseReferenceForPcn(pcnCaseReference);
-            if (existingEnforcement != null && existingEnforcement != event.caseReference()) {
+            if (existingEnforcement != null && existingEnforcement.longValue() != event.caseReference()) {
                 throw new IllegalArgumentException(
                     "PCN case " + pcnCaseReference
                         + " is already linked to enforcement case " + existingEnforcement
